services/api/requirements.txt
# 데이터 처리
pandas>=2.0
numpy>=1.24
pydantic>=2.0

# 예측 모델
scipy>=1.10
scikit-learn>=1.3
lightgbm>=4.0

# HTTP / 환경변수
httpx>=0.25
requests>=2.28
requests-cache>=1.2
retry-requests>=2.0
python-dotenv>=1.0
PyYAML>=6.0

# 날씨 API
openmeteo-requests>=1.2

# Supabase 연동
supabase>=2.0

# FastAPI 서버
fastapi>=0.110
uvicorn[standard]>=0.27

# ECMWF GRIB(data_pipeline/ecmwf_forecast_data.py) — 선택 설치, Windows 는 eccodes 까다로울 수 있음
# ecmwf-opendata>=0.3.0
# ecmwf-api-client>=1.6.0
# xarray>=2024.1.0
# cfgrib>=0.9.10
